% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_records.R
\name{clean_records}
\alias{clean_records}
\title{Creating a summary file of counts of germplasm, herbarium and records for a given species using a CSV file   (Pre-analysis function)}
\usage{
clean_records(Workspace, species_csv, species_list, run_version)
}
\arguments{
\item{Workspace}{A forder where the pipeline will be executed}

\item{species_csv}{CSV file name located at /Workspace/parameters/input to be splitted}

\item{species_list}{A list of species to be analyzed}

\item{run_version}{The version of the analysis used (e.g 'v1')}
}
\value{
It returns a raster file saved in /Workspace/parameters/input to be used to calculate ex-situ conservation indicators
}
\description{
Creates a summary file of counts of germplasm, herbarium as well split records
 in csv files per species to allow calculate all indicators provided by GapAnalysisR R package
}
\examples{
clean_records(Workspace,species_csv,species_list,run_version)

 dir  <-  'E:/CIAT/workspace/Workspace_test'
 run_version  <- 'v1'
 species_csv <- 'Cucurbita_CWR_2019_09_30.csv'

 x <- clean_records(Workspace,species_csv,species_list,run_version)

}
\references{
Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
