% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffers.R
\name{create_buffers}
\alias{create_buffers}
\title{Creating buffer around germplasm accessions extracted from a CSV file (Pre-analysis function)}
\usage{
create_buffers(species, Workspace, buff_dist, run_version)
}
\arguments{
\item{species}{A name species compiled using '_'  to call occurrences files from Workspace/parameter/occurrences folder}

\item{Workspace}{A forder where the pipeline will be executed}

\item{buff_dist}{Distance around germplasm accessions geographical coordinates in km^2}

\item{run_version}{The version of the analysis used (e.g 'v1')}
}
\value{
It returns a raster file saved in /Workspace/parameters/input to be used to calculate ex-situ conservation indicators
}
\description{
Creates a raster using circular buffer around germplasm accessions for a given species using a CSV file
 and the R packages fasterize and geobuffer
}
\examples{
create_buffers(species,Workspace,buff_dist,run_version)

 dir  <-  'E:/CIAT/workspace/Workspace_test'
 run_version  <- 'v1'
 species_list <- c('Cucurbita_cordata',
  'Cucurbita_digitata',
  'Cucurbita_foetidissima',
  'Cucurbita_palmata')

 x <- create_buffers(species,Workspace,buff_dist,run_version)

}
\references{
Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
