% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folder_structure.R
\name{create_folder_structure}
\alias{create_folder_structure}
\title{Creating folder and subfolder structure for a species list extracted from a CSV file  (Pre-analysis function)}
\usage{
create_folder_structure(species_list, dir, run_version, initial,
  create_sp_dirs)
}
\arguments{
\item{species_list}{A list of species to be analyzed when a species list is provided}

\item{dir}{A forder to run the gap analysis including inputs and outputs}

\item{run_version}{The version of the analysis used (e.g 'v1')}

\item{initial}{(logical) The function  will create the initial folder structure for the analysis}

\item{create_sp_dirs}{(logical) The function  will create folder for each species outputs in /gap_anaylsis}
}
\value{
It returns a folder structure with the following subfolder schema:
\tabular{lll}{
  \tab workspace \tab species name as species parameter \cr
  \tab └───parameters  \tab #Folder where the inputs are hosted \cr
  \tab │   ├───raster \tab # Subfolder to add mask, protected areas and ecoregions shapefile \cr
  \tab │   ├───occurrences  \tab #Folder where a CSV file is used to filter species accessions occurrences \cr
  \tab └───gap_analysis \tab #Folder where results will be saved if create_sp_dirs=T\cr
}

 Sources used for analysis were obtained from Khoury et al., 2019 using the following download links:

\tabular{lll}{
\tab Protected area  \tab http://dx.doi.org/10.17632/2jxj4k32m2.1#file-fc8ffd3f-8d79-481b-9ccf-4cf19b8480ec \cr
\tab World mask \tab http://dx.doi.org/10.17632/2jxj4k32m2.1#folder-1b61c978-5753-4ab6-bbc7-11f8e4b062cf \cr
\tab Terrestrial ecoregions \tab http://maps.tnc.org/files/shp/terr-ecoregions-TNC.zip \cr
}
}
\description{
Creates the output folders structure for each species to host results obtained by the gap analysis pipeline
 This function is used internally in the create_folder_structure function when an initial species list is provided
}
\examples{
create_folder_structure(species_list,dir,run_version,initial=F,create_sp_dirs=T)

 dir  <-  'E:/CIAT/workspace/Workspace_test'
 run_version  <- 'v1'
 species_list <- c('Cucurbita_cordata',
  'Cucurbita_digitata',
  'Cucurbita_foetidissima',
  'Cucurbita_palmata')

 x <- create_folder_structure(species,dir,run_version,initial=F,create_sp_dirs=F)

}
\references{
Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Wiersema, J. H. (2019).
Data for the calculation of an indicator of the comprehensiveness of conservation of useful wild plants.
Data in Brief, 22, 90–97. https://doi.org/10.1016/j.dib.2018.11.125
}
