% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sp_dirs.R
\name{create_sp_dirs}
\alias{create_sp_dirs}
\title{Creating folder and subfolder structure of gap analysis results for a given species  (Pre-analysis function)}
\usage{
create_sp_dirs(species, gap_dir, run_version)
}
\arguments{
\item{species}{A name species compiled using '_'  to call occurrences files from Workspace/parameter/occurrences folder}

\item{run_version}{The version of the analysis used (e.g 'v1')}

\item{Workspace}{A forder where the pipeline will be executed}
}
\value{
It returns a folder structure with the following subfolder schema:
\tabular{lll}{
  \tab species name \tab species name as species parameter \cr
  \tab └───v1  \tab #Run version \cr
  \tab │   ├───gap_analysis \tab # Gap analysis results \cr
  \tab │   ├───combined  \tab #Folder where results of fcs_combine.R are saved \cr
  \tab │   ├───exsitu  \tab #Folder where Ex-situ indicators results will be saved \cr
  \tab │   └───insitu \tab #Folder where In-situ indicators results will be saved \cr
  \tab └───redList \tab #Folder where results of eooAoo function will be saved \cr
}
}
\description{
Creates the output folders structure for each species to host results obtained by the gap analysis pipeline
 This function is used internally in the create_folder_structure function
}
\examples{
create_sp_dirs('Cucurbita_digitata',Workspace,'v1')

Workspace  <-  'E:/CIAT/workspace/Workspace_test/workspace'
run_version  <- 'v1'
species_list <- c('Cucurbita_cordata',
 'Cucurbita_digitata',
 'Cucurbita_foetidissima',
 'Cucurbita_palmata')

 run_version <-'v1'
lapply(1:length(species_list),function(i){
   species <- species_list[[i]]
   x <- create_sp_dirs(species,Workspace,run_version)
   print(paste0(species,' DONE!'))
})

}
\references{
Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
