% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eooAoo.R
\name{eooAoo}
\alias{eooAoo}
\title{Area of occupancy (AOO) and extent of occurrence (EOO) in km2 (Auxiliar functions)}
\usage{
eooAoo(species, Workspace, run_version)
}
\arguments{
\item{species}{A name species compiled using '_'  to call occurrences files from Workspace/parameter/occurrences folder}

\item{Workspace}{A forder where the pipeline will be executed}

\item{run_version}{The version of the analysis used (e.g 'v1')}
}
\value{
It returns a data frame file saved at gap_analysis folder with six columns:

\tabular{lcc}{
 taxon \tab Species name \cr
 EOO Area km2 \tab Area of occupancy in km^2 (AOO) for a given species \cr
 EOO Status \tab IUCN conservation using (AOO)  \cr
 AOO \tab extent of occurrence (EOO) in km2 for a given species \cr
 AOO adjusted Minimum \tab Minimum adjusted polygon for AOO for a given species \cr
 AOO Status \tab IUCN conservation using (AOO) \cr
}
}
\description{
Calculates Area of occupancy (AOO) Area and extent of occurrence (EOO) in km2
 using a species distribution model, an occurrences CSV file and categorize the species conservation status according IUCN parameters
}
\examples{
eooAoo('Cucurbita_digitata',Workspace,'v1')

Workspace  <-  'E:/CIAT/workspace/Workspace_test/workspace'
run_version  <- 'v1'
species_list <- c('Cucurbita_cordata',
 'Cucurbita_digitata',
 'Cucurbita_foetidissima',
 'Cucurbita_palmata')

 run_version <-'v1'
lapply(1:length(species_list),function(i){
   species <- species_list[[i]]
   x <- eooAoo(species,Workspace,run_version)
   print(paste0(species,' DONE!'))
})

}
\author{
Dan Carver
}
