% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_combine.R
\name{fcs_combine}
\alias{fcs_combine}
\title{Combining Ex-situ and In-situ gap analysis results in one comprehensive score}
\usage{
fcs_combine(species, Workspace, run_version)
}
\arguments{
\item{species}{A name species compiled using '_'  to call occurrences files from Workspace/parameter/occurrences folder}

\item{Workspace}{A forder where the pipeline will be executed}

\item{run_version}{The version of the analysis used (e.g 'v1')}
}
\value{
It returns a data frame file saved at gap_analysis folder with nive fields:

\tabular{lcc}{
 ID \tab Species name \cr
 FCSex \tab Ex-situ final conservation score \cr
 FCSin \tab In-situ final conservation score \cr
 FCSc_min \tab Final conservation score mininum value between FCSin and FCSex \cr
 FCSc_max \tab Final conservation score maximum value between FCSin and FCSex \cr
 FCSc_mean \tab Final conservation score average value between FCSin and FCSex \cr
 FCSc_min_class \tab Final conservation category using  FCSc_min value \cr
 FCSc_max_class \tab Final conservation category using  FCSc_max value \cr
 FCSc_mean_class \tab Final conservation category using  FCSc_mean value \cr
}
}
\description{
Concatenates the SRSin, GRSin, ERSin, SRSex, GRSex, ERSex and Final conservation scores
 in one unique CSV file with a combined conservation score using the minimun, average and maximum values between FCSin and FCSex scores
 and provides final conservation categories based on the scores above mentioned.
}
\examples{
fcs_combine('Cucurbita_digitata',Workspace,'v1')

Workspace  <-  'E:/CIAT/workspace/Workspace_test/workspace'
run_version  <- 'v1'
species_list <- c('Cucurbita_cordata',
 'Cucurbita_digitata',
 'Cucurbita_foetidissima',
 'Cucurbita_palmata')

 run_version <-'v1'
lapply(1:length(species_list),function(i){
   species <- species_list[[i]]
   x <- fcs_combine(species,Workspace,run_version)
   print(paste0(species,' DONE!'))
})

}
\references{
Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
