% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sp_names.R
\name{get_sp_names}
\alias{get_sp_names}
\title{Obtaining species names from a CSV file}
\usage{
get_sp_names(Workspace, species_csv)
}
\arguments{
\item{Workspace}{A forder where the pipeline will be executed}

\item{species_csv}{A CSV file name located at /Workspace/parameters/input to be splitted}
}
\value{
It returns a species list in an object
}
\description{
Creates a species list reading a CSV file to use in the gap_analysis pipeline
}
\examples{
clean_records(Workspace,species_csv)

 Workspace = "E:/CIAT/workspace/Workspace_test/workspace"
 species_csv <- 'Cucurbita_CWR_2019_09_30.csv'

 x <- get_sp_names(Workspace,species_csv)

}
\references{
Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
