% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRSex.R
\name{grs_exsitu}
\alias{grs_exsitu}
\title{Germplasm representativeness score estimation (Ex-situ conservation indicators).}
\usage{
grs_exsitu(species, Workspace, run_version)
}
\arguments{
\item{species}{A name species compiled using '_'  to call the raster files (SDM and germplasm buffer)
from Workspace/parameters/inputs folder and occurrences files from Workspace/parameter/occurrences folder}

\item{Workspace}{A forder where the pipeline will be executed}

\item{run_version}{The version of the analysis used (e.g 'v1')}
}
\value{
It returns a data frame file saved at gap_analysis folder with four columns:

\tabular{lcc}{
 ID \tab Species name \cr
 SPP_AREA_km2 \tab Area occupied by the species using as input a SDM thresholded file in tiff format \cr
 G_AREA_km2 \tab Area occupied by the germplasm accessions in a species distribution model \cr
 GRS \tab GRSex result \cr
}
}
\description{
Performs an estimation of germplasm representativeness score for ex-situ gap analysis (GRSex) using Ramirez-Villegas et al., (2010) methodology
This function uses a germplasm buffer raster file (e.g. CA50) and a thresholded species distribution model.
 \deqn{GRSex = min(100,(Germplasm buffer area/species distribution in protected areas)*100)}
}
\examples{
grs_exsitu('Cucurbita_digitata',Workspace,'v1')

Workspace  <-  'E:/CIAT/workspace/Workspace_test/workspace'
run_version  <- 'v1'
species_list <- c('Cucurbita_cordata',
 'Cucurbita_digitata',
 'Cucurbita_foetidissima',
 'Cucurbita_palmata')

 run_version <-'v1'
lapply(1:length(species_list),function(i){
   species <- species_list[[i]]
   x <- grs_exsitu(species,Workspace,run_version)
   print(paste0(species,' DONE!'))
})

}
\references{
Ramírez-Villegas, J., Khoury, C., Jarvis, A., Debouck, D. G., & Guarino, L. (2010).
A Gap Analysis Methodology for Collecting Crop Genepools: A Case Study with Phaseolus Beans.
PLOS ONE, 5(10), e13497. Retrieved from https://doi.org/10.1371/journal.pone.0013497

Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
