% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_code.R
\name{Integrate_gap_analysis_function}
\alias{Integrate_gap_analysis_function}
\title{Run a combined gap analysis combining all the functions of GapAnalysR package}
\usage{
Integrate_gap_analysis_function(dir, run_version, Workspace, species,
  buff_dist, species_csv, pre_analysis, insitu, exsitu, eooAoo)
}
\arguments{
\item{dir}{A forder to run the gap analysis including inputs and outputs}

\item{run_version}{The version of the analysis used (e.g 'v1')}

\item{Workspace}{A forder where the pipeline will be executed}

\item{species}{Species name provided by user or the get_sp_names function}

\item{buff_dist}{Distance around germplasm accessions geographical coordinates in km^2}

\item{species_csv}{CSV file name located at /Workspace/parameters/input to be splitted}

\item{pre_analysis}{(logical) This option allows to run the set of pre-analysis functions:
\tabular{lll}{
 create_buffers \tab  \cr
}}

\item{insitu}{(logical) This option allows to calculate In-situ indicators:

\tabular{lll}{
 srs_insitu \tab  \cr
 grs_insitu \tab  \cr
 ers_insitu \tab  \cr
 fcs_insitu \tab  \cr
}}

\item{exsitu}{(logical) This option allows to calculate Ex-situ indicators:
\tabular{lll}{
 srs_exsitu \tab  \cr
 grs_exsitu \tab  \cr
 ers_exsitu \tab  \cr
 fcs_exsitu \tab  \cr
 gap_map_exsitu \tab  \cr
}}

\item{eooAoo}{(logical) This option allows to calculate EOO and AOO to get IUCN conservation status}
}
\value{
This function run the complete GapAnalysisR pipeline in one unique function
}
\description{
Allows calculate all indicators provided by GapAnalysisR R package.
 Please use the Pre_analysis_function function before run this pipeline
}
\examples{
Integrate_gap_analysis_function(dir,run_version,Workspace,buff_dist,pre_analysis,insitu,exsitu,eooAoo)

 dir  <-  'E:/CIAT/workspace/Workspace_test'
 run_version  <- 'v1'
 species_csv <- 'Cucurbita_CWR_2019_09_30.csv'
 Workspace  <-  'E:/CIAT/workspace/Workspace_test/workspace'
 buff_dist <- 50000 #50km radius
 pre_analysis <- T
 insitu <- T
 exsitu <- T
 eooAoo <- T

 Integrate_gap_analysis_function(dir,run_version,Workspace,species,buff_dist,species_csv,pre_analysis,insitu,exsitu,eooAoo)
 #INCLUDE INPUT FILES BEFORE RUN THE NEXT FUNCTION

 species_list <- get_sp_names(Workspace,species_csv)

 lapply(1:length(species_list),function(i){
  species <- species_list[[i]]
  x <- Integrate_gap_analysis_function(dir,run_version,Workspace,species,buff_dist,species_csv,pre_analysis,insitu,exsitu,eooAoo)
  cat(paste0(species,"| status:",x),"\\n")
 })

}
\references{
Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
