% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pre_analysis_function.R
\name{Pre_analysis_function}
\alias{Pre_analysis_function}
\title{Run the first steps for the gap analysis}
\usage{
Pre_analysis_function(dir, species_csv, run_version)
}
\arguments{
\item{dir}{A forder to run the gap analysis including inputs and outputs}

\item{species_csv}{CSV file name located at /Workspace/parameters/input to be splitted}

\item{run_version}{The version of the analysis used (e.g 'v1')}

\item{Workspace}{A forder where the pipeline will be executed}
}
\value{
This function run the first steps for the GapAnalysisR pipeline
}
\description{
Creates initial steps for the gap analysis as folder structure, species folder results and species list names
}
\examples{
Pre_analysis_function(dir,run_version,Workspace,buff_dist,pre_analysis,insitu,exsitu,eooAoo)

 dir  <-  'E:/CIAT/workspace/Workspace_test'
 run_version  <- 'v1'
 species_csv <- 'Cucurbita_CWR_2019_09_30.csv'
 Workspace  <-  'E:/CIAT/workspace/Workspace_test/workspace'

x <- Pre_analysis_function(dir,species_csv,run_version,Workspace)

}
\references{
Khoury, C. K., Amariles, D., Soto, J. S., Diaz, M. V., Sotelo, S., Sosa, C. C., … Jarvis, A. (2019).
Comprehensiveness of conservation of useful wild plants: An operational indicator for biodiversity
and sustainable development targets. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2018.11.016
}
